1.完善账号状态、账号逻辑删除相关逻辑;2.管理端添加管理员模块

+++ pblog-intf/src/main/java/com/pblog/common/result/PageResult.java
@@ -18,7 +18,7 @@ import java.util.List;
 @NoArgsConstructor
 public class PageResult implements Serializable {
 
-    private long total; //总记录数
+    private Integer total; //总记录数
 
     private List records; //当前页数据集合
